refactor: start new file structure migration
Creates directories for new file structure, adds css files in each directory, and moves test files to the appropriate directories
